+++ llampc/llampc/rollout/rk6.py
@@ -50,7 +50,7 @@ def odeintRK4_batch(bank_params, diffequation, h, known_params, x0, u):
         k3 = h * step(b_p, x0 + k2 / 2)
         k4 = h * step(b_p, x0 + k3)
         result = x0 + (k1 + 2 * k2 + 2 * k3 + k4) / 6
-        return result.astype(jnp.float16)
+        return result.astype(jnp.float32)
 
     return jax.vmap(rk4)(bank_params)
